simplify binding authorizer
The old code would loop through all bound claims and for each matching (group/resource) claim it
would try to find the first (!) matching exported claim. And then check the verbs. If no exported
claim is found, the set of allowed verbs is empty, so the HasAny() check can never succeed. Hence
if no exported claim exists, we do not even have to check the verbs.

This commit improves the loop by comparing all bound claims against all (!) exported claims (with
the matching GRI) and skipping any guaranteed-false checks if no exported claim is found.

This loop now is ready to be extended to check for references, too.

On-behalf-of: @SAP christoph.mewes@sap.com