fix concurrent map read access and write panic
The controller already uses the RWMutex to guard the calls to state.apply() (which is important because
we have 2 worker threads in this controller), but failed to also guard the state.getGVKRs() call.
apply() would write to the underlying map in the DefaultRESTMapper, getGVKRs() would read. Boom.

This panic happens extremely rarely and only surfaced when I moved the initialisation of the
DiscoveringDynamicSharedInformerFactory up from server.NewServer() into server.NewConfig(). With
that change, kcp begun to panic in maybe 30% of cases within the first 5 seconds of runtime.

Since this is in fact a race condition, this fix still make sense even when the panic barely ever
surfaces right now.

On-behalf-of: @SAP christoph.mewes@sap.com